feat: require-await
Disallow functions that are `async` but do not use await. This builds
off of the `no-return-await` by removing another unneeded promise wrap.

Warning: this will not catch functions that return two types, ie.
promise / or empty array. In this case, the developer will have to be
concious of the return type and use try+catch to handle the promise
error and utilize await.