feat(ts): explicit boundary types
I want to propose using this rule to enforce explicitly typing functions
that are exported at a boundary. The aim here is to force
changes to actual types instead of implicit types - resulting in a more
cohesive typing environment. Changes to functions with explict types are
more safe, because you cannot accidently change the type without visible
fallout.

This does not affect local functions, which are absolutely capable of
using implicit typing, and should be encouraged to do so.