feat(common): prefer-destructuring
Rule details: https://eslint.org/docs/rules/prefer-destructuring

This is to prevent the proliferation of `const a = array[0]`, when the
preferred syntax `const [a] = array` is available. I've expressed the
sane defaults that the rule has available, in case we want to change the
way it behaves for objects and arrays individually.