feat: add generator star spacing rules
This removes the deprecated `no-spaced-func` rule and replaces it with `func-call-spacing` which supports
generators better.

Valid code with these rules:

```js
async function* () {}
async function() {}
```

Code not valid with these rules:

```js
async function * () {}
async function*() {}
async function () {}
```

I think this pattern allows for the best readability, and allows for easy identification of the
`*` in the generator signatures.